ajustes de codigo
se realizo una modificación el lo visual y cambio de carpetas
