+++ admin/css/style.css
@@ -1,0 +1,94 @@
+@import url('https://fonts.googleapis.com/css2?family=Cardo&family=Montserrat:wght@300&display=swap');
+*{
+    padding: 0;
+    margin: 0;
+}
+body{
+    display: grid;
+    grid-template-columns: 1fr 4fr;
+}
+aside{
+    background: rgb(246, 236, 249);
+    min-height: 100vh;
+    border-right: 1px solid rgba(207, 204, 207, 0.856);
+    padding: 10px 0 10px 10px ;
+}
+.logo{
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 7px;
+}
+.logo img{
+    border-radius: 100px;
+}
+
+.logo h1{
+    font-size: 1.3rem;
+    line-height: 1.4rem;
+}
+.menu{
+    margin-top: 60px;
+}
+.menu ul{
+    list-style: none;
+}
+.menu ul a{
+    text-decoration: none;
+    font-size: 600;
+    color: #333;
+    display: flex;
+    align-items: center;
+    gap: 10px;
+    padding: 15px ;
+    cursor: pointer;
+}
+.menu ul a:hover{
+    background: #333;
+    color: #fff;
+}
+main{
+    padding: 40px;
+}
+.header{
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    margin-bottom: 50px;
+}
+.header h2{
+    font-size: 1.8rem;
+}
+
+.btnaction{
+    background: #555;
+    padding: 10px 20px;
+    color: #fff;
+    border-radius: 4px;
+    text-decoration: none;
+}
+.btnaction:hover{
+    background: #f0f0f0;
+    color: #555;
+}
+table{
+    min-width: 100%;
+}
+table thead{
+    background: #555;
+    color: #fff;
+}
+table thead th,table tbody td{
+    padding: 8px;
+    text-align: left;
+}
+table tbody tr:nth-child(even){
+    background: #f3f3f3;
+}
+#insert_data{
+    display: none;
+}
+
+#btnBorrar {
+    display: none;
+}
