--- anime-unid/admin/css/style.css
@@ -1,114 +1,0 @@
-* {
-    padding: 0%;
-    margin: 0%;
-}
-
-body {
-    display: grid;
-    grid-template-columns: 1fr 4fr;
-}
-
-aside {
-    background: rgb(246, +246, +249);
-    min-height: 100vh;
-    border-right: 1px solid rgb(220, +220, +228);
-    padding: 10px 0 10px 10px;
-}
-
-.logo {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 7px;
-}
-
-.logo img {
-    border-radius: 50%;
-}
-
-.logo h1 {
-    font-size: 1.3rem;
-    line-height: 1.4rem;
-}
-
-.menu {
-    margin-top: 60px;
-}
-
-.menu ul {
-    list-style: none;
-}
-
-.menu ul a {
-    text-decoration: none;
-    font-weight: 600;
-    color: #333;
-    display: block;
-    align-items: center;
-    gap: 10px;
-    padding: 15px;
-    cursor: pointer;
-}
-
-.menu ul a i {
-    font-size: 1.2rem;
-}
-
-.menu ul a:hover {
-    background: #333;
-    color: #fff;
-}
-
-main {
-    padding: 40px;
-}
-
-.header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin-bottom: 50px;
-}
-
-.header h2 {
-    font-size: 1.8rem;
-}
-
-.btnActions {
-    background: #555;
-    padding: 10px 20px;
-    color: #fff;
-    border-radius: 4px;
-    text-decoration: none;
-}
-
-.btnActions:hover {
-    background: #f0f0f0;
-    color: #555;
-}
-
-table {
-    min-width: 100%;
-}
-
-table thead {
-    background: #333;
-}
-
-table thead th,
-table thead td {
-    padding: 5px;
-    text-align: left;
-}
-
-table tbody tr:nth-child(even) {
-    background: #f0f0f0;
-}
-
-#insert_data {
-    display: none;
-}
-
-#btnBorrar {
-    display: none;
-}
